pyproject.toml
[project]
name = "quadruped-mpc-genesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "casadi>=3.7.2",
    "genesis-world>=0.4.6",
    "numpy",
    "pin>=3.9.0",
    "torch>=2.11.0",
]
